LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.clearRect.nonfinite.worker-expected.txt


PASS clearRect() with Infinity/NaN is ignored 
